fuse: replace DirEntryList.FixMode()
As pointed out by @slackner at
https://github.com/hanwen/go-fuse/commit/8581b7e950b2d520370eba8997ddd9493192e9c7#r35845728
FixMode() failed to take into account the length
of the file name and the padding.

To fix this without adding a lot of additional
pointer arithmetics, have AddDirLookupEntry() return
pointers to both EntryOut and _Dirent. This way,
the caller can easily modify _Dirent.Typ, if needed.

Change-Id: I2fc7336561ec639a7eba517a1c4a04e7679413d4